pyproject.toml
[project]
name = "luxical-tw"
authors = [
    { name = "Luke Merrick", email = "luke@datology.ai" },
    { name = "DatologyAI Team" },
]
dynamic = ["version"]
description = "Lexical dense embeddings. Fast like light (lux). Traditional Chinese Fork."
readme = "README.md"
requires-python = ">=3.11,<3.14"
classifiers = [
    "Programming Language :: Python :: Implementation :: CPython",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
license = "Apache-2.0"
license-files = ["LICENSE"]
dependencies = [
    "arrow-tokenize",
    "numba>=0.61.2",
    "scipy>=1.15.3",
    "torch>=2.5.0",
    "tqdm>=4.65.0",
    "uniplot>=0.21.2",
    "s3fs>=2024.9.0",
    "fsspec>=2024.9.0",
    "opencc-python-reimplemented>=0.1.7",
    "jieba-fast-dat>=0.58",
    "fastembed>=0.7.4",
]

[project.optional-dependencies]
sklearn = [
    "scikit-learn>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/luxical_tw/__about__.py"

[tool.uv.sources]
"arrow-tokenize" = { workspace = true }

[tool.uv.workspace]
members = ["arrow_tokenize"]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "pytest>=8.4.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.4",
    "transformers[torch]>=4.51.3",
    "twine>=6.2.0",
]
misc = [
    "datasets>=3.6.0",
    "fasttext>=0.9.3",
    "matplotlib>=3.10.3",
    "scikit-learn>=1.7.0",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "transformers[torch]>=4.51.3",
    "ipywidgets>=8.1.7",
    "jupyterlab>=4.4.3",
]

[tool.pyright]
enableTypeIgnoreComments = true
typeCheckingMode = "standard"
include = [
    "src",
    "tests",
]
exclude = [
    "projects",
    "data",
    "build",
    "dist",
    ".venv",
    "arrow_tokenize/target",
    "arrow_tokenize/dist",
]
venvPath = "."
venv = ".venv"
reportMissingModuleSource = "none"
reportMissingImports = "none"
reportPrivateImportUsage = "none"
reportOptionalSubscript = "warning"
reportAttributeAccessIssue = "warning"
